Completed part 4 of the tutorial
Scenes are now loaded correctly based on the number of players in the game. Since we are not using a lobby the scene is reloaded when a player joing to match the number of players with the scene. If more than the max number of players join the new player is correctly put in a new scene.